test: run round-trip over configCases marked roundTrip instead of a separate suite
Re-bundle the output of `configCases` opted in via `test.config.js` { roundTrip: true }
(RoundTripConfigCases) rather than maintaining duplicate fixtures under roundTripCases/.
