--- test/RoundTripTestCases.basictest.js
@@ -1,5 +1,0 @@
-"use strict";
-
-const { describeCases } = require("./RoundTripTestCases.template");
-
-describeCases({ name: "RoundTripTestCases" });
