+++ test/configCases/asset-modules/url-module-analyzable/test.config.js
@@ -1,0 +1,7 @@
+"use strict";
+
+// Re-bundle the emitted output through webpack (see RoundTripConfigCases): the
+// analyzable `new URL("./file.png", import.meta.url)` must survive the round trip.
+module.exports = {
+	roundTrip: true
+};
