--- test/roundTripCases/url/import-meta-url/index.js
@@ -1,1 +1,0 @@
-export const logo = new URL("./logo.png", import.meta.url);
