--- test/roundTripCases/url/import-meta-url/logo.png
@@ -1,1 +1,0 @@
-ROUND-TRIP-PNG
